Change BtnC to Menu and make separate component
