+++ src/components/Mypage/Menu.tsx
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+
+import { theme } from "styles/theme";
+
+import ReportGmailerrorredRoundedIcon from "@mui/icons-material/ReportGmailerrorredRounded";
+import HelpOutlineRoundedIcon from "@material-ui/icons/HelpOutlineRounded";
+import AssignmentOutlinedIcon from "@material-ui/icons/AssignmentOutlined";
+import PortraitRoundedIcon from "@material-ui/icons/PortraitRounded";
+import ExitToAppRoundedIcon from "@material-ui/icons/ExitToAppRounded";
+import KeyboardArrowLeftRoundedIcon from "@material-ui/icons/KeyboardArrowLeftRounded";
+
+type MenuProps = {
+  icon: string;
+  onClick: () => void;
+  children: React.ReactElement;
+};
+
+const getIcon = (icon: string) => {
+  const styleIcon = {
+    ...theme.font15_icon,
+    marginRight: "8px",
+  };
+  switch (icon) {
+    case "report":
+      return <ReportGmailerrorredRoundedIcon style={styleIcon} />;
+    case "ask":
+      return <HelpOutlineRoundedIcon style={styleIcon} />;
+    case "policy":
+      return <AssignmentOutlinedIcon style={styleIcon} />;
+    case "credit":
+      return <PortraitRoundedIcon style={styleIcon} />;
+    case "logout":
+      return <ExitToAppRoundedIcon style={styleIcon} />;
+  }
+};
+
+const Menu = (props: MenuProps) => {
+  const [isHover, setHover] = useState(false);
+  const style = {
+    display: "flex",
+    alignItems: "center",
+    width: "fit-content",
+    color: isHover ? theme.purple : undefined,
+  };
+  const styleText = {
+    ...theme.font14,
+    color: "inherit",
+  };
+  return (
+    <div
+      style={style}
+      className="BTNC"
+      onMouseEnter={() => setHover(true)}
+      onMouseLeave={() => setHover(false)}
+      onTouchStart={() => setHover(true)}
+      onTouchEnd={() => setHover(false)}
+      onClick={() => props.onClick()}
+    >
+      {getIcon(props.icon)}
+      <div style={styleText}>{props.children}</div>
+      {isHover && <KeyboardArrowLeftRoundedIcon style={theme.font15_icon} />}
+    </div>
+  );
+};
+
+export default Menu;
